Make APEX install optional. Improve sym link creation, comments and network files.

+++ database/ol7_122/scripts/start.sh
@@ -1,5 +1,7 @@
-# Handle shutdowns
-# docker stop --time=30 {container}
+echo "******************************************************************************"
+echo "Handle shutdowns." `date`
+echo "docker stop --time=30 {container}" `date`
+echo "******************************************************************************"
 function gracefulshutdown {
   dbshut $ORACLE_HOME
 }
@@ -8,22 +10,42 @@ trap gracefulshutdown SIGINT
 trap gracefulshutdown SIGTERM
 trap gracefulshutdown SIGKILL
 
-# Fixes the config using the contents of the volume.
-# Necessary when using persistent volume as "rm" and "run" will reset the config
-# under the ORACLE_HOME.
+echo "******************************************************************************"
+echo "Define fixConfig function." `date`
+echo "Fixes the config using the contents of the volume." `date`
+echo "Necessary when using persistent volume as "rm" and "run" will reset the config" `date`
+echo "under the ORACLE_HOME." `date`
+echo "******************************************************************************"
 function fixConfig {
-  cp -f /u02/config/oratab /etc/oratab 
-  ln -s /u02/config/${ORACLE_SID}/orapw${ORACLE_SID} ${ORACLE_HOME}/dbs/orapw${ORACLE_SID}
-  ln -s /u02/config/${ORACLE_SID}/spfile${ORACLE_SID}.ora ${ORACLE_HOME}/dbs/spfile${ORACLE_SID}.ora
-  ln -s /u02/config/${ORACLE_SID}/admin ${ORACLE_BASE}/admin
-  ln -s /u02/config/${ORACLE_SID}/fast_recovery_area ${ORACLE_BASE}/fast_recovery_area
-  rm -Rf ${ORACLE_BASE}/diag
-  ln -s /u02/config/${ORACLE_SID}/diag ${ORACLE_BASE}/diag
+  cp -f /u02/config/oratab /etc/oratab
+  if [ ! -L ${ORACLE_HOME}/dbs/orapw${ORACLE_SID} ]; then
+    ln -s /u02/config/${ORACLE_SID}/orapw${ORACLE_SID} ${ORACLE_HOME}/dbs/orapw${ORACLE_SID}
+  fi
+  if [ ! -L ${ORACLE_HOME}/dbs/spfile${ORACLE_SID}.ora ]; then
+    ln -s /u02/config/${ORACLE_SID}/spfile${ORACLE_SID}.ora ${ORACLE_HOME}/dbs/spfile${ORACLE_SID}.ora
+  fi
+  if [ ! -L ${ORACLE_BASE}/admin ]; then
+    ln -s /u02/config/${ORACLE_SID}/admin ${ORACLE_BASE}/admin
+  fi
+  if [ ! -L ${ORACLE_BASE}/fast_recovery_area ]; then
+    ln -s /u02/config/${ORACLE_SID}/fast_recovery_area ${ORACLE_BASE}/fast_recovery_area
+  fi
+  if [ ! -L ${ORACLE_BASE}/diag ]; then
+    rm -Rf ${ORACLE_BASE}/diag
+    ln -s /u02/config/${ORACLE_SID}/diag ${ORACLE_BASE}/diag
+  fi
 }
 
-# Create a listener.ora file if it doesn't already exist.
+echo "******************************************************************************"
+echo "Create networking files if they don't already exist." `date`
+echo "******************************************************************************"
 if [ ! -f ${ORACLE_HOME}/network/admin/listener.ora ]; then
-echo "LISTENER = 
+  echo "******************************************************************************"
+  echo "First start, so create networking files." `date`
+  echo "******************************************************************************"
+
+  cat > ${ORACLE_HOME}/network/admin/listener.ora <<EOF
+LISTENER = 
 (DESCRIPTION_LIST = 
   (DESCRIPTION = 
     (ADDRESS = (PROTOCOL = IPC)(KEY = EXTPROC1)) 
@@ -31,10 +53,39 @@ echo "LISTENER =
   ) 
 ) 
 USE_SID_AS_SERVICE_listener=on
-" > ${ORACLE_HOME}/network/admin/listener.ora
+INBOUND_CONNECT_TIMEOUT_LISTENER=400
+EOF
+
+  cat > ${ORACLE_HOME}/network/admin/tnsnames.ora <<EOF
+LISTENER = (ADDRESS = (PROTOCOL = TCP)(HOST = 0.0.0.0)(PORT = 1521))
+
+${ORACLE_SID}= 
+(DESCRIPTION = 
+  (ADDRESS = (PROTOCOL = TCP)(HOST = 0.0.0.0)(PORT = 1521))
+  (CONNECT_DATA =
+    (SERVER = DEDICATED)
+    (SERVICE_NAME = ${ORACLE_SID})
+  )
+)
+
+${PDB_NAME}= 
+(DESCRIPTION = 
+  (ADDRESS = (PROTOCOL = TCP)(HOST = 0.0.0.0)(PORT = 1521))
+  (CONNECT_DATA =
+    (SERVER = DEDICATED)
+    (SERVICE_NAME = ${PDB_NAME})
+  )
+)
+EOF
+
+  cat > ${ORACLE_HOME}/network/admin/sqlnet.ora <<EOF
+SQLNET.INBOUND_CONNECT_TIMEOUT=400
+EOF
 fi
 
-# Check if database already exists.
+echo "******************************************************************************"
+echo "Check if database already exists." `date`
+echo "******************************************************************************"
 if [ ! -d /u02/oradata/${ORACLE_SID} ]; then
 
   # The database files don't exist, so create a new database.
@@ -59,19 +110,25 @@ if [ ! -d /u02/oradata/${ORACLE_SID} ]; then
     -emConfiguration NONE                                                      \
     -ignorePreReqs
 
-  # Set the PDB to auto-start.
+  echo "******************************************************************************"
+  echo "Set the PDB to auto-start." `date`
+  echo "******************************************************************************"
   sqlplus / as sysdba <<EOF
 alter system set db_create_file_dest='/u02/oradata';
 alter pluggable database ${PDB_NAME} save state;
 exit;
 EOF
 
-  # Store config files in case persistent volume is used.
+  echo "******************************************************************************"
+  echo "Store config files in case persistent volume is used." `date`
+  echo "******************************************************************************"
   dbshut ${ORACLE_HOME}
   mkdir -p /u02/config/${ORACLE_SID}
     
   cp /etc/oratab /u02/config/
-  # Flip the auto-start flag.
+  echo "******************************************************************************"
+  echo "Flip the auto-start flag." `date`
+  echo "******************************************************************************"
   sed -i -e "s|${ORACLE_SID}:${ORACLE_HOME}:N|${ORACLE_SID}:${ORACLE_HOME}:Y|g" /u02/config/oratab
   cp -f /u02/config/oratab /etc/oratab 
     
@@ -85,10 +142,13 @@ EOF
   fixConfig;
   dbstart ${ORACLE_HOME}
 
-  # Install APEX.
-  cd ${ORACLE_HOME}/apex
+  if [ "$INSTALL_APEX" = "true" ]; then
+    echo "******************************************************************************"
+    echo "Install APEX." `date`
+    echo "******************************************************************************"
+    cd ${ORACLE_HOME}/apex
 
-  sqlplus / as sysdba <<EOF
+    sqlplus / as sysdba <<EOF
 alter session set container = ${PDB_NAME};
 create tablespace apex datafile size 1m autoextend on next 1m;
 @apexins.sql APEX APEX TEMP /i/
@@ -114,18 +174,22 @@ alter user APEX_PUBLIC_USER identified by ${APEX_PASSWORD} account unlock;
 
 exit;
 EOF
-
+  fi
 else
 
-  # The database already exists, so start it.
+  echo "******************************************************************************"
+  echo "The database already exists, so start it." `date`
+  echo "******************************************************************************"
   fixConfig;
 
   dbstart $ORACLE_HOME
 
 fi
 
-# Tail the alert log file as a background process
-# and wait on the process so script never ends.
+echo "******************************************************************************"
+echo "Tail the alert log file as a background process" `date`
+echo "and wait on the process so script never ends." `date`
+echo "******************************************************************************"
 tail -f ${ORACLE_BASE}/diag/rdbms/${ORACLE_SID}/${ORACLE_SID}/trace/alert_${ORACLE_SID}.log &
 bgPID=$!
 wait $bgPID
